Authentication-Authorization

+++ controllers/review.js
@@ -1,16 +1,20 @@
 const Listings = require("../models/listing");
+const Review = require("../models/review");
 const Reviews = require("../models/review");
+const ExpressError = require("../utils/ExpressError");
 
 module.exports.createReview = async (req, res) => {
   const listing = await Listings.findById(req.params.id);
+
   const newReview = new Reviews(req.body.review);
+  console.log(req.user.id);
+  newReview.author = req.user.id;
 
   listing.reviews.push(newReview);
 
   await newReview.save();
   await listing.save();
 
-  console.log(listing);
   console.log("New Review Saved!");
 
   res.json({
@@ -22,14 +26,20 @@ module.exports.createReview = async (req, res) => {
 module.exports.deleteReview = async (req, res) => {
   const { id, reviewId } = req.params;
 
-  const listing = await Listings.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
-  const review = await Reviews.findByIdAndDelete(reviewId);
+  const review = await Reviews.findById(reviewId);
 
-  console.log(listing);
+  if (!review.author.equals(req.user.id)) {
+    return res.status(403).json({
+      message: "You are not authorized to delete this review",
+    });
+  }
+
+  await Listings.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
+  await Reviews.findByIdAndDelete(reviewId);
 
   res.json({
     status: true,
-    message: "success",
+    message: "Review deleted successfully",
     reviewData: review,
   });
 };
